FalsePosition Function
This was an assignment to create a function to solve for the roots of a function using the FalsePosition method. The purpose of the assignment was to familiarize the way in which false position works as well as get feel for how a function file works in matlab and how to properly code it.